refactor GameScoreBoard to its own folder and files
