+++ src/Components/Game/Components/GameScoreBoard/GameScoreBoard.css
@@ -1,0 +1,4 @@
+.game-score-board {
+    width: 100%;
+    display: flex;
+}
